Update The Deprecated; Remove The Unused
Many of the endpoints in the Plaid API have changed or been removed since
this library was written. Update the methods accordingly. Remove methods
for endpoints that no longer exist, Add methods for endpoints that are new,
rename things that should be renamed, update params and variables where
appropriate. Do all the things!